requirements.txt
kafka-python==2.0.2
msgpack==1.1.0
numpy==2.1.2
pandas==2.2.3
py4j==0.10.9.7
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
